# swaggers/contents/swagger.yml
openapi: "3.0.2"
info:
    title: Contents API
    version: "0.1"
    description: 영상/교재/위키피디아/아티클 API
servers:
  - url: http://192.168.153.173:8080/
    description: example server address
...
paths:       
...
  /videos:
    get:
      tags:
        - Video
      summary : 비회원 영상 리스트 조회
      description : 비회원 영상 리스트 조회
      parameters:
      - name: curriculumId
        in: query
        description: 영상 주제
        required: false
        schema:
          type: integer
          format: integer
      - name: sort
        in: query
        description: DATE(날짜순)/HIT(조회순)
        required: false
        schema:
          type: string
          format: string
          default: DATE
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                $ref: '#/components/schemas/VideoListModel'
  /users/{user_id}/videos/{video_id}:
    get:
      tags:
        - Video
      summary : 유저의 영상 조회
      description : 유저의 북마크 정보를 포함한 영상 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: video_id
        in: path
        description: 영상 아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                $ref: '#/components/schemas/VideoModel'
  /users/{user_id}/videos:
    get:
      tags:
        - Video
      summary : 유저의 영상 리스트 조회
      description : 유저의 북마크 정보를 포함한 영상 리스트 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: curriculumId
        in: query
        description: 영상 주제
        required: false
        schema:
          type: integer
          format: integer
      - name: sort
        in: query
        description: DATE(날짜순)/HIT(조회순)
        required: false
        schema:
          type: string
          format: string
          default: DATE
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                $ref: '#/components/schemas/VideoListModel'
  /users/{user_id}/videos/recommend:
    get:
      tags:
        - Video
      summary: 유저의 추천 영상 조회
      description: 유저의 추천 받은 영상 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: curriculumId
        in: query
        description: 영상 주제
        required: false
        schema:
          type: integer
          format: integer
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  recommended:
                    type: boolean
                  recommends:
                    $ref: '#/components/schemas/VideoListModel'
  /users/{user_id}/videos/bookmark:
    get:
      tags:
        - Video
      summary : 유저의 북마크 영상 리스트 조회
      description : 유저가 북마크 선택한 영상 리스트 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: curriculumId
        in: query
        description: 영상 주제
        required: false
        schema:
          type: integer
          format: integer
      - name: sort
        in: query
        description: DATE(날짜순)/HIT(조회순)
        required: false
        schema:
          type: string
          format: string
          default: DATE
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                $ref: '#/components/schemas/VideoListModel'
  /users/{user_id}/videos/{video_id}/bookmark:
    put:
      tags:
        - Video
      summary : 유저의 영상 북마크 추가
      description : 유저가 영상에 북마크 추가
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: video_id
        in: path
        description: 영상 아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
    delete:
      tags:
        - Video
      summary : 유저의 영상 북마크 삭제
      description : 유저가 영상에 북마크 삭제
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: video_id
        in: path
        description: 영상 아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true                               
  /videos/{video_id}/hit:
    post:
      tags:
      - Video
      summary: 영상 조회수 업데이트
      description: 해당 영상의 조회수를 증가시키고 LRS 저장
      parameters:
      - name: video_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true                            
  /users/{user_id}/videos/{video_id}/hit:
    post:
      tags:
      - Video
      summary: 영상 조회수 업데이트
      description: 해당 영상의 조회수를 증가시키고 LRS 저장
      parameters:
      - name: user_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: string
          format: string
      - name: video_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true                  
  /users/{user_id}/videos/{video_id}/quit:
    post:
      tags:
      - Video
      summary: 영상 quit 업데이트
      description: 해당 영상의 quit LRS 저장
      parameters:
      - name: user_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: string
          format: string
      - name: video_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: integer
          format: integer
      - name: duration
        in: query
        description: 컨텐츠 소비 시간
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
...
  /stats/contents/byDate:
    get:
      tags:
      - Statistics
      summary: 날짜별 컨텐츠 소비 조회
      description: 날짜별 컨텐츠 소비 조회
      parameters:
      - name: dateFrom
        in: query
        description: 시작 날짜
        required: true
        schema:
          type: string
          format: date
      - name: dateTo
        in: query
        description: 마지막 날짜
        required: true
        schema:
          type: string
          format: date
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  size:
                    type: integer
                  stats:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                        hit:
                          type: object
                          properties:
                            total:
                              type: integer
                            video:
                              type: integer
                            book:
                              type: integer
                            wiki:
                              type: integer
                            article:
                              type: integer        
  /stats/contents/byUsers:
    get:
      tags:
      - Statistics
      summary: 유저 컨텐츠 소비 조회
      description: 유저 컨텐츠 소비 조회
      parameters:
      - name: userIds
        in: query
        description: 유저아이디 리스트
        required: true
        schema:
          type: array
          items:
            type: string
          format: array
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  size:
                    type: integer
                  stats:
                    type: array
                    items:
                      type: object
                      properties:
                        userId:
                          type: string
                          format: string
                        hit:
                          type: object
                          properties:
                            total:
                              type: integer
                            video:
                              type: integer
                            book:
                              type: integer
                            wiki:
                              type: integer
                            article:
                              type: integer
                        bookMark:
                          type: object
                          properties:
                            total:
                              type: integer
                            video:
                              type: integer
                            book:
                              type: integer
                            wiki:
                              type: integer
                            article:
                              type: integer
components:
  schemas:
    VideoModel:
      type: object
      properties:
        videoId:
          type: integer
        videoSrc:
          type: string
        title: 
          type: string
        imgSrc: 
          type: string
        subject: 
          type: string
        totalTime: 
          type: number
        startTime:
          type: number
        endTime:
          type: number
        hit: 
          type: integer
        bookmark: 
          type: boolean
        videoType:
          type: string
          description: YOUTBUE(유튜브 영상)/SELF(자체 제작)
        uks:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
    VideoListModel:    
      type: object
      properties:
        size: 
          type: integer
        videos:
          allOf:
          - $ref: '#/components/schemas/VideoModel'